Less_Cache: Remove impossible file_exists check for Less_Parser
Due to the order of operations, this check was unfortunately performing
file I/O on every call even when the class was already loaded.

This was added in 2014 with commit 42100fd0. I don't know what scenario
this was for, but these classes are nowadays either already loaded
or automatically loaded by either the Composer autoloader, or our
own standalone Less_Autoloader.