Use __DIR__ where possible and optimise standalone Less_Autoloader
It is common in autoloaders to call file_exists. This is generally
done so that if the prefix is fairly generic (and has classes in
multiple repositories/packages) then each one can check its directory
to see if it has it. E.g. if example/foo.git and example/bar.git each
register namespace Example and then define a single class in that
namespace.

In the case of LESS, this autoloader already decides that it is the
final owner of LESS_ since it has an else branch that throws a fatal
error.

As such, remove the file_exists check and require it directly. This
allows full use of opcache and removes the need for file I/O. If it
doesn't exist, PHP will emit an appropriate erorr instead.