test: Minor cleanup and focus more on lessjs/less.php diffs
I couldn't get all of this to work, which seems mainly due to logic
relating to comparing to previous runs. I've disabled this for now
in favour of focussing it more on unresolved differences between
lessjs and less.php. Anything that's passing is ideally added to
the PHPUnit test to avoid regressions, we don't need the manual
comparison page for that.