Less_Functions: Fix "Undefined property" warning on bad minmax arg
To reproduce before this patch:

    $ echo '.foo { color: min(rgb(1,1,1)); }' | bin/lessc  -
    PHP Warning:  Undefined property: Less_Tree_Color::$value
    in /less.php/lib/Less/Functions.php:685

The min() and max() handlers were lacking the property_exists check
that other LESS function handlers already performed. The output is
the same either way since it previously yielded implicit null which
is also not an array, but the check ensures we do so without a
PHP warning. Phan still needs additional convincing.

Change-Id: Ibc55c58537be3a25caa57f34e340463da57c0521