Misc clean up around Less_Ruleset for readability and less.js parity
In prep for fixing T204816, I've uploaded cosmetic changes and
documenting of findings along the way separately, to keep the next
diff cleaner and easier to review, leaving no doubt about what changes
were needed to fix the bug.

The only non-comment change here is the removal of the instanceof
check in Ruleset->compile, which improved parameter typehinting
showed to be redundant. And indeed, it doesn't exist in the equivalent
less.js code.

The check was added in 2013 with commit c13e32398, which at the time
aimed for less.js 1.3.3 parity, but that didn't have the check
either [1], and more importantly both back then and now, the same
Ruleset->compile method already has other conditional access to
the $env object, such as `$env->shiftFrame();` a few lines down.

[1] https://github.com/less/less.js/blob/v1.3.3/dist/less-1.3.3.js#L3283-L3313

Bug: T204816
Change-Id: Ic1aa57c8c410514b02022869ce24826c8991a06f