Less_Parser: Only support `cache_method=serialized`
== Background ==

less.php comes with two caching features, both of which are opt-in,
and neither of which MediaWiki makes use of (MediaWiki instead caches
the whole CSS output directly in Memcached/php-apcu, and invokes
less.php effectively as a lambda only, with no internal cache)

1. Less_Cache - This is what our documentation recommends, and can
   cache the whole CSS output after parsing a Less input file.
   This is opt-in only, by using Less_Cache instead of Less_Parser.

2. Less_Parser 'cache_dir' - This is barely documented and is not used
   by default unless called via Less_Cache. It is represents an
   incremental cache that stores PHP objects that allow partial
   re-use of a Less_Tree during a cache miss.

This patch affects the latter, incremental cache.

== What ==

Having a built-in cache feature is fine, but the one that we have is
far more complex and involved that I'm comfortable with supporting
or maintaining. I'd like to start by reducing it to a single
implementation, namely the `serialized` method that stores serialized
PHP objects in text files.

The other two cache methods rely on generating executable PHP code,
including an attempt to reverse-engineer class invocations, which add
complexity and "NewObj" indirection for all instantiations. The next
commit will remove all this (for improved Phan coverage, PhpStorm
automation, and performance).

It also seems problematic from a security perspectice to generate,
store, and access, PHP executables, plus impact on limited
php-opcache size (PHP 7.0+), and no obvious on-going purge mechanism
in-place, I'd rather support the text format only.

Change-Id: If6745bdc37e2ca78ef45437cb76260b9a6729bad