Less_Environment: Treat `url(#myid)` as absolute URL
When the 'relativeUrls=false' option is used or when 'rootpath' is
known, as MediaWiki does by default, then image references are
rewritten to be relative to a given URL. This allows the resulting CSS
to be safely served from a different URL. For example, you can bundle
it with other CSS files into /w/load.php or /static.123.

A bare `#hash` URL refers to something in the DOM. Match behaviour
of less.js to fix this bug.

Note that this doesn't involve a fixture file as neither we nor
upstream currently have fixtures that parse with non-default options.

The way I checked upstream behaviour:
* Run `php -S localhost:4000`
* Open <http://localhost:4000/test/?dir=lessjs&file=comments>
* In browser console:
  `await less.render('div { background: url(img.svg#id); }', { rootpath: '/x/' });`
  `await less.render('div { background: url(#myid); }', { rootpath: '/x/' });`
  which yields:
  "div {\n  background: url(/x/img.svg#id);\n}\n"
  "div {\n  background: url(#myid);\n}\n"

Bug: T331649
Change-Id: I7ce5214d3f281a4673c0a2afabc7acbb498ab3c7