Fix "Undefined property" warning when calc() is used with CSS var()
== Less_Tree_Operation ==

In Less_Tree_Operation::operate(), we call Dimension::operate() with
argument $b (named $other there), which Dimension expects to be a
number with a unit, i.e. something we can perform math on.

Change Operation::operate() to not call this if argument $b is
something else, which avoids the reported PHP warning of $other->unit
being undefined. Noting that the only class with a 'unit' property
is Less_Tree_Dimension.

As a side-effect, this immediately fixes another bug, namely that the
"var()" right-hand side of `calc(100% - var(foo))` was completely
ignored previously when it wasn't a number. For example, it can be
a CSS keyword or CSS function like var() that the browser handles
client-side. Noting that CSS also natively supports math. LESS should
only handle math when the input uses PHP variables or literals, and
thus effectively has a constant outcome.

Previously, Dimention::operate() would implicitly read $other->value
as NULL when $other value wasn't a numerical Dimension object, and
thus the math operation silently failed before this commit, e.g.:
> "100% - var(foo)"
> "100 - NULL" (as %)
> "100%"

With the warning avoided/fixed, we now correctly preserve the
equation in its entirely when $b is e.g. a var() call.

== Less_Tree_Call ==

The above change on its own is insufficient, as that elevates the
warning to a fatal error. The added test case exposes this,
where `calc(100% - var(foo) - var(bar))` is parsed as:

  Call (calc)
  \-- Expression
      \-- Operation (-)
          |-- Operation (-_
          |    |- Dimension (100%)
          |    |- Call (var foo)
          |-- Call (var bar)

The above change, makes the first Operation pair no longer compile
Dimension+Call down to a Dimension, instead preserving Operation as
literal CSS output. This means the second Operation will now see a
left-hand value of Operation, which doesn't have an "operate" method,
and thus correctly triggers the `throw` statement that I cleaned up,
because that's not something we can compute server-side.

Fix this by recognising `calc()` higher up as something that
server-side LESS math should not be performed on in the first place.

This effectively backports https://github.com/less/less.js/pull/3162
(https://github.com/less/less.js/commit/a48c24c4dd3c13e00a20ece803237)
for the similarly reported issue upstream at
https://github.com/less/less.js/issues/974.

That bugfix was originally released with less.js 3.0.0, but the
part of that I'm backporting here doesn't change behaviour of any
of our less.js 2.5.3 compliance tests, thus safe to backport.

We already threw a fatal error when one of the arguments wasn't a
Dimension or Color (no other class has the 'operate' method), so this
only adds non-fatal outcomes to previously fatal outcomes.

Bug: T331688
Change-Id: Idd443a76372682de4edddeb64ab5a65b23bbd3ce