build: Move RedundantVarName rule from `exclude` to `severity`
This should avoid the problem [1] with Libraryupgrader, which is configured
to try to "fix" any disabled rules automatically, and if it manages to
do so without leaving any warnings, it will propose removing that rule
as part of routine package upgrades. The "severity" mechanism is a stronger
signal that it will not try to fix.

[1] https://gerrit.wikimedia.org/r/c/mediawiki/libs/less.php/+/897780

Bug: T278474
Change-Id: Ib2e96220162514843c362e0ec6140c5852b6c5ec