test: Move FunctionTest data to data/ subdirectory
Move these out of the "Fixtures" directory as they are not
fixtures used by the FixturesTest in a general input/output kind
of way.

Instead, the files under functions/ are related to a specific
configuration as done by FunctionTest.

This is in prep for a new test suite where I'll use the same
data/ directory for other one off files.

Change-Id: I7dfb3af2e3cbb42b45acd7053a6109ccca1d8237