MixinCall,Selector: Fix memory leak and selector bug
There were two failures here, the first one is a memory leak.

I copied the mixins.less to a separate input.less file, and reduced
it until I had a small reproducible example that caused only the
memory leak. This was:

```
.clearfix() {
	color: blue;
}
.clearfix {
	.clearfix();
}
.clearfix {
	.clearfix();
}
```

Using git-blame to find when this part of the test was added upstream,
we find https://github.com/less/less.js/commit/8afeae92d, which reveals
a very small change to mixin.Call in JS, that indeed is missing from
our PHP port. Adding the $originalRuleset override, fixed the memory
leak.

After that, there was another test failure revealed in mixins.less:

> 1) phpunit_FixturesTest::testFixture with data set
>    "Fixtures/lessjs-2.5.3 mixins" (…)
> Failed asserting that two strings are identical.
> --- Expected
> +++ Actual
> @@ @@
>  #container {
>    color: black;
>    border: 1px solid black;
> +  border: do not match me;
>    border-color: orange;
>    background-color: grey;
>  }

Tracing this with git-blame brings us to
https://github.com/less/less.js/commit/62eb799b40, which makes a
small change to Selector.CacheElements that indeed was missing from
our port. Applying that change, fixed this test failure.

Bug: T352829
Change-Id: Ief17d5e6b9c494c31981e107153a0c55ad37eb77