Less_Functions: Fix `replace()` when passed multiple replacements
I2fcb32e5eb (29cf2b99d3) fixed a bug in the `replace` function where
the absence of the 'g' modified was ignored, and all occurrences were
always replaced. However, in fixing this bug, it created a new one:
when the 'g' modified was present, the replacement was run twice,
first as a single replace and then as a multiple replace.
This caused all sorts of buggy output.

Fix this by running the replacement only once, either with the limit
set to 1 (if 'g' is not set) or with the limit not set (if 'g' is set).

Bug: T358631
Change-Id: Ica9de4b14385167ad300ad05e1c3995af0ad8ade