test: Rename 'expected' to 'override' and remove redundant copies
The css/ directory contains the upstream specification.

The expected/ directory is a duplicate with minor differences where
we've accepted a difference where we believe the test is useful to
enable even if it is not yet completely passing (instead of disabling
the test entirely in FixturesTest::KNOWN_FAILURE).

To make this easier to maintain, let's keep only the files that are
different, and update FixtureTest.php to compare against css/ by
default, unless an override/ file exists.

The name "override" is probably clearer this way, instead of
"expected" which did not make it obvious since css/ is also the
expected output.

Change-Id: I1f358a359dea839c880f96c85041cbcd6a32aef2