test: Remove unversioned test/Fixtures/lessjs/ directory
This unversioned directory was a mixture of various different less.js
1.x and 2.x specs from upstream. In T349580, we added the latest
Less.js 2.x spec from Less.js 2.5.3 at test/Fixtures/lessjs-2.5.3/
which is what PHPUnit and compare.php use now.

The last remaining reference was from bench.php where I had included
one of its files (strings.less). Preserve this for benchmark continuity
in a new bench-strings directory, and rename it to reflect that it
came from
<https://github.com/less/less.js/blob/v1.5.0/test/less/strings.less>

The second benchmark was unnamed, but largely about Bootstrap with
a few small files thrown in as well. This was already unstable since
the directories it used do get changed sometimes (e.g. less.php/ and
lessjs/), so remove those from it in one last changed and rename it
to more clearly be about a pinned version.

Change-Id: I715358f7c284e0e4905c18aa053d6590610beb42