test: Enable bootstrap 3.0.3, 3.1, and 3.2 in CI. Passing!
The reason we were failing with some minor cosmetic differences is
that the "css" files were generated with a much older version of
Less.js. I've now re-generated them with the reference version of
Less.js 2.x, and low and behold, we fully pass, not even with minor
cosmetic differences in float precision or whitespace.

I regenerated them using Less 2.7 instead of 2.5 because 2.5 was
not compatible with any version of Node.js I could install. It
did not make a difference for the output, and the fact that we're
passing 2.7 already is if anything better than passing 2.5.

While at it, also:
* remove "expected" overrides, since these now unused.
* remove "bootstrap3" fixture, was already unused.
* remove "bootstrap-3.1-sourcemap" fixture, was already unused.

Change-Id: I3aca8b90efdf1791bd84af497ffc39bb44bbc8d8