Less_Parser: Faster matching by inlining matcher() chains
The matcher() method was a utility custom to Less.php.
It adds indireciton and overhead that makes the parser slower,
and it also made static analysis by Phan less useful because
these don't look like normal function calls, so the return value
is not understood by Phan.

Since the introduction of expectChar(), it is also no longer needed
for matcher() to handle '#' has a way to handle single characters,
this logic is now unused.

After inlining the mathcer() chains, the only remaining use of it
is expect(). Move the code for matchReg() and $this->$tok() there
directly, same as upstream less-2.5.3.js.

Change-Id: I017c1a6726d9d7a98f0d427c239ef83020e27343