Less_Parser: Update matchRegex() to less-2.5.3.js string/array return
I like methods to have consistent return values, but, unfortunately
upstream implements `parserInput.$re` such that it returns the string
from match[0] for patterns that no capture groups, and returns the
match array whole for patterns that do have capture groups.

As much as that's different from what I'd do, I do often make mistakes
in porting because I forget a match[0] in our code. For consistency,
match the upstream code more exactly so that porting becomes easier.

One benefit from this change is that it makes chaining possible with
like `parseFoo() ?? matchChar ?? matchRegex() ?? parseBar()`. Today,
this would produce an array return value if the regex matches whereas
it would return a string for the others. In the next patch, I will
make use of this.

Change-Id: I2be5f78cf312ddde1da6a3131523bbfffc89fb25