Fix ParseError for comments after rule name or in `@keyframes`
This commit ports required logic from an upstream patch
for Less.js 1.7:
https://github.com/less/less.js/commit/e045362c901e6bb5

This simply eats up comments and is easy to backport.
Less.js 2.5.3 actually takes a different approach, it uses a
ParserInput class and stores comments in a commentStore.

This first change will allow the parser to eat up comments
and prevent syntax errors. As we don't store the comments yet
the generated CSS won't contain all comments because we have
no way to re-inject those later, yet. The comments.css override
allows us to make sure that we properly compile this Less
without errors.

The comments2.less tests comments between rule names
and block definitions. Those are not only parsed properly,
but also should be put in the generated CSS.

What Less.js 2.0 does, they parse comments when handling
spaces in skipWhitespace, instead of manually consuming comments
in specific places during parsing, and comments are pushed to an
array. Then, they either empty the comments array,
or get the first one and add it to the output.

In T353132 I'm going to do a follow-up on this work and
match Less.js 2.0 implementation - introducing the ParserInput
class with commentStore. As ParserInput change is going
to affect overall parsing, I prefer to do it in a separate
task.

Bug: T353131
Change-Id: I95db673e90726909eaa108870ee600fa4eca351e