test: Merge Fixtures/bug-reports/ into Fixtures/less.php/
Motivation: Reduce noise from skipped tests that are not actionable.

== Background ==

These are not upstream less.js bug reports, but our own bug reports
from when less.php was maintained by oyejorge. This means we have some
freedom over changing them, because ultimately upstream takes
precedence.

Over time, it seems some of these are no longer passing and we have
disabled them as precaution. But it is time to take a closer look at
these and decide to remove or keep them, or to file a regression task
if we believe there is a bug still.

Also, it seems a number of these fixtures were added in master many
years ago, with the intention to change the behaviour, but then the
next step never happened, e.g. because the bug turned out to be invalid
(and the failing test case left behind) or because upstream less.js
indicated that the behaviour was working as expected.

* For all cases, rename them to `oyejorge-` to clarify that these
  refer to bug reports from that repo, to make those easier to find
  in case we need more information.

* For all cases, rename them to include a word or two about what
  they are testing (based on reading the bug report) so that if we
  do need to make minor changes to these, we are less likely to
  accidentally defeat the purpose of the regression test.

== Clarify ==

* 305 - Rename loose "import.less" to "305-import-grantparent" so
  that its reason for existing can be discovered, and to prevent it
  being re-used for other random tests.

== Remove ==

These were pre-emptively committed but declined or invalid:

* 110 - https://github.com/oyejorge/less.php/issues/110
  This seemingly never passed, and re-purposed internally
  the fixture of a test for issue 80. The behaviour of url
  rewriting via imports is already covered via other tests.

* 259 - https://github.com/oyejorge/less.php/issues/259
  This was declined, but the fixture seemingly added with
  an intent to make it work one day, but this was declined
  by upstream Less.js as expected.
  https://github.com/less/less.js/issues/2704

* 280 - https://github.com/oyejorge/less.php/issues/208
  The ordering of `@charset` similarly seems to have
  never passed and the issue remained open / was declined
  upstream. There are numerous other tests for this already,
  so we don't need this as our own custom test.

Change-Id: I92f05a9d73c01ecbaaf02b79db26147b93279671