Less_Functions: Avoid clobbering `clamp()` with internal helper
Upstream defines clamp() as a private helper function. It is not
exposed on the `colorFunctions` object, and thus not an actual
LESS feature, callable by LESS source files. I.e. it does not take
Less_Tree objects, is not documented on https://lesscss.org/functions/,
and there are no upstream tests for it. It doesn't exist.

Our code accidentally exposed this helper function as a public
method, thus causing Less_Tree_Call to call it during compilation,
instead of leaving it unchanged, as native CSS built-in.

This wasn't an issue an practice for the other helper functions, because
they don't happen to share a name with something that exists in CSS.
This coincidence seems to be limited to clamp(), however for good
measure, I am fixing the other helper functions as well, by making them
private as well, same as upstream Less.js 2.5.3.

* Rename from clamp() to _clamp(), to prevent the name clash and fix
  the example input from T363728. This does not prevent exposure to
  Less_Tree_Call, since this change by itself mean anyone can
  write "_clamp()" in a `.less` file, and still call the same internal
  function, which is undesirable.

  The underscore is analogous to how upstream implements helpers as
  locally scoped functions, thus in a "namespace" that is separate
  from the public functions that it implements as keys in a plain
  object.

* Make helper function natively private, so that we're confident that
  it cannot be called from elsewhere. This prevents it from executing,
  but does not prevent Less_Tree_Call from trying (and failing),
  since it uses method_exists() instead of is_callable(). Change that
  at the same time, so that we do not expose and vary public behaviour
  at all by existence of private helpers.

  This also means that the names of internal helpers become avaiable
  for public use via `Less_Parser->registerFunction()`, same as
  upstream. Add a test for this to prevent regressions.

Also:

* Update f1.less fixture, used by FunctionTest.php, which we added
  in 1ede985b45ce2ab3b24b318e86a87c097fe1b4ee. I did not realize at
  the time that `number()` was not a real function. I simply picked it
  as the first function in Less_Functions that has an error case that
  I could cover as example. Change the error test to rgb() instead.

Closes https://github.com/wikimedia/less.php/issues/78.

Bug: T363728
Change-Id: I9811afd09070066b45f33851a6ebba6b7830f17b