test: Enable lessjs-3.13.1/compression fixture
Same as lessjs-2.5.3/compression before. For simple NameValue rules,
we can't apply the `0.1` -> `.1` compression. For the optimum compression,
we use CSSMin afterwards anyway, so this isn't important.

The test is still useful to ensure the rest of the output works,
the parser option supported and pass through everywhere, and uses the
correct spacing, matching upstream.

Bug: T288498
Change-Id: Iadf8da2ee33fa6906a8c2c7d3f8231d21a38130a