Improve support for preserving `!important` via variables
Implement !important scope handling in Less_Tree_Rule by introducing
the Environment::$importantScope variable. By returning new Rulee
with evaluated value instead of self we additionally solved
variable scoping.

Values for px, pt and pc had to be adjusted (higher precision) to
avoid small miscalcualtions when converting units.

We had to introduce a new temporary variable in Parser called
`$cachedEvaldRules`. Before this change, `Less_Tree_Rule::compile()`
had side effets - it was overriding the `Less_Tree_Rule::$value`
with compiled value. The `Parser::getVariables()` was using
`$this->rules`, and because values was replaced in-place - it was
able to retrieve all compiled variables.
With this change, the `Less_Tree_Rule` doesn't have side effects,
therefore the `Parser::getCss()` didn't store the evaluated rules.

For now let's cache the evaluated Nodes in `Parser::$cachedEvaldRules`
but it would be great to revisit this approach in the near future.

This PR also makes the lessjs 3.11.3 `variables.less` pass.

Bug: T362341
Change-Id: I05e784d1af2d04f9b892b5c543c017c62defec74