Remove outdated `Less.php.combine` file
This was used in the past to create a single-file distribution, such
as wp-less-to-css does at
https://github.com/bassjobsen/jbst/tree/2.1.0/vendor/wp-less-to-css/less-php

To my knowledge, no other user of this package uses this, and the
above has not been updated since 10 years ago, when this package was
still under the name oyejorge/less.php version 1.6.1.

If and when they update, whether this is a standalone file or a
standalone directory will be a trivial difference. They can use either
the Composer package, or they can commit the library as standalone
subdirectory. In that case they would only need to change the require_once
from "Less.php" to "Less/Autoloader.php", and then use it the same
as before.

Change-Id: I9be2468d21e185f875d76125ac89199395530d26