test: Reduce verbose details from skipped tests
In d09158300 we converted known failures from ignored to "skipped"
tests. However, this has the downside of adding one or two pages
of output at the end of the `composer test` output, thus making
the failure details harder to find.

Disable verbose=true by default so that this is reduced to a skipped
count. You can run it with -v or --verbose to display each one in
elaborate detail.

Change-Id: Iae2508470b3358e6d8186a62bcd3a319599adaf6