Less_Tree_Color: Remove internal logic for $isTransparentKeyword
== Why ==

This code was added in Less.php 1.8.1 (f62c0fba84, and 1d1283167f)
back in 2013, based on code from Less.js 1.5 [1], which indeed
implemented support for preserving the original form of "transparent"
with this exact logic.

However, this was replaced upstream in Less.js 2.0 [2] with the
general "originalForm" parameter.

In Less.php, we ported "originalForm" in I0e6fa0fa8 (79ae252b5c,
T352866) last year, but we left behind the now-redundant
$isTransparentKeyword logic, which Less.js 3.13 does not have,
so we should not have it either.

[1]: https://github.com/less/less.js/commit/862d7ed0ab9d42995a268e07714868ac85e48366
[2]: https://github.com/less/less.js/commit/3322609de5fe0ff6fc3d125ed6df7393a9ab0484

== What ==

This patch does a full review of Less_Tree_Color compared to
less-3.13.1.js#Color.prototype, and syncs all minor differences
that clean up the code and don't change behaviour.

Any major differences I've marked with TODO, as those would change
behaviour and add new features, which requires more careful review
and testing, which can we done later.

Bug: T352866
Change-Id: Ifac4b328470caa36b74263d5e09e549e0f5f952e