Less_Parser: Use matching types when resetting internals in unsetInput()
Setting properties to null would require documentation to include
nullable, but then Phan would want null check all over the file,
which we don't want. We can satisfy the types instead.

These nulls are never seen in practice because unsetInput() is the
last thing we do during GetRules(), until the next time it is called
which starts with setInput() and sets all of these correctly.

Change-Id: I6be6db137f9a29c481dc008cbb50526307109adb