build: Add missing documentation to public class properties
Add doc-typehints to protected class properties found by the
PropertyDocumentation sniff to improve the documentation.

Enable the sniff to avoids that new code is missing type declarations.
This is focused on documentation and does not change code.

In some situation phan is wrong or confused with the new added
documentation, add @phan-var and move suppression around

Change-Id: Icbd9a1574b7929a23afdeb8ae1b68836a1d88517