Implement isVisble for Less_Tree_Anonymous
We have made the nodeVisible property to be slightly different than that of less.js.

Less.js defaults to null but we can't default to null because we don't have a full implementation of the visibility feature.

So for now we would just assume that all nodes defaults to true and implement the full visibility feature in a later task.

Bug: T393383
Change-Id: I5b203b24fd208781a25b3551be535263464eeb57