Add support for CSS4 `#RRGGBBAA` colors
This fixes hex colors compiling to invalid syntax when using 4-char
or 8-char notations with an alpha channel.

A major change in this patch was matching Less.js 3.13 output of hsl
and hsla functions in CSS output, instead of converting to hex or rgb
format.

All browsers we support, including Grade C browsers, support this as they have been part of browsers for a long time.
See https://caniuse.com/?search=hsla

Bug: T403056
Change-Id: I3022d677f2f99ccb2f78e105928ae178acc41f68