Fix interpolation of string with number followed by underscore
This was fixed upstream in Less.js v2.6.0. I caught this during
a review of the nearby parseEntitiesDimension as part of T403056.

Remove the unneeded "@" silence operator from the ord() call.
This was added in 2013 with commit 51e29cbf to deal with null
when the position has reached outside the bounds of $this->input.
A better fix was added in Ie8d5da3ed4 (67807cb7) using `?? ''`,
which removes the need for error silencing, because it is now always
passed a string.

Ref https://github.com/less/less.js/issues/2462
Ref https://github.com/oyejorge/less.php/issues/293

Change-Id: I968056b7e45c9752b8ce3833979729ac8230dc6d