Less_ImportVisitor: Fix "Using null as an array offset" PHP 8.5 warning
In less.js, they do something like "fullPath in importVisitor.recursionDetector" which is kind of the equivalent to array_key_exists. We would also still encounter this error if we use array_key_exists because less.js also passes in null to the function. These values are set in ImportManager.prototype.push and if you narrow down, the actual callback is ImportVisitor.prototype.onImported.

Bug: T411400
Change-Id: I4ad03834f87e01ae316494fab1f8205480722f46