Less_Parser: Fix SetImportDirs() type hint to accept callable without key
The way this is used in practice, in MediaWiki, in ParserTest::testSetImportDirs,
and in the parameter documentation on this very method, is without a
(meaningless) string key.

During a downstream PR at https://github.com/flarum/framework/pull/4225/files,
I found PHPStan failing on:

```
$parser->SetImportDirs([ $function ]);

// PHPStan:
// Parameter #1 $dirs of method Less_Parser::SetImportDirs()
// expects
// array<string, (callable(): mixed)|string>,
// array{callable(): mixed}
// given.
```

Note that PHPStan 1.11.0 understands `phan-param` as alias for `param`,
so it will parse it the same way, which is good, this isn't an issue
with Phan vs PHPStan. Our documented typehint is actually incomplete.

Ref https://github.com/phpstan/phpstan/discussions/10996

Change-Id: Id5953eaa3e3a62f6b29ec533a5393ef32dbb1098